effecthash: clean up split impls into new crate
