+++ crates/core/component/dex/src/swap/action.rs
@@ -2,6 +2,7 @@ use anyhow::Context;
 use ark_ff::Zero;
 use decaf377::Fr;
 use penumbra_asset::{balance, Balance, Value};
+use penumbra_effecthash::{EffectHash, EffectingData};
 use penumbra_num::Amount;
 use penumbra_proto::{
     core::component::dex::v1alpha1 as pbc, penumbra::core::component::dex::v1alpha1 as pb,
@@ -36,6 +37,14 @@ impl Swap {
         let commitment_input_2 = input_2.commit(Fr::zero());
 
         commitment_input_1 + commitment_input_2 + self.body.fee_commitment
+    }
+}
+
+impl EffectingData for Swap {
+    fn effect_hash(&self) -> EffectHash {
+        // The effecting data is in the body of the swap, so we can
+        // just use hash the proto-encoding of the body.
+        self.body.effect_hash()
     }
 }
 
@@ -79,6 +88,12 @@ pub struct Body {
     pub delta_2_i: Amount,
     pub fee_commitment: balance::Commitment,
     pub payload: SwapPayload,
+}
+
+impl EffectingData for Body {
+    fn effect_hash(&self) -> EffectHash {
+        EffectHash::from_proto_effecting_data(&self.to_proto())
+    }
 }
 
 impl DomainType for Body {
