+++ crates/core/component/governance/Cargo.toml
@@ -46,6 +46,7 @@ penumbra-distributions = { path = "../distributions", default-features = false }
 penumbra-chain = { path = "../chain", default-features = false }
 penumbra-asset = { path = "../../../core/asset", default-features = false }
 penumbra-keys = { path = "../../../core/keys", default-features = false }
+penumbra-effecthash = { path = "../../../core/effecthash", default-features = false }
 penumbra-num = { path = "../../../core/num", default-features = false }
 
 # Penumbra dependencies
