+++ crates/core/component/governance/src/proposal.rs
@@ -29,6 +29,15 @@ pub struct Proposal {
     pub payload: ProposalPayload,
 }
 
+/*
+// TODO: is this needed?
+impl EffectingData for Proposal {
+    fn effect_hash(&self) -> EffectHash {
+        EffectHash::from_proto_effecting_data(&self.to_proto())
+    }
+}
+ */
+
 /// The protobuf type URL for a transaction plan.
 pub const TRANSACTION_PLAN_TYPE_URL: &str = "/penumbra.core.transaction.v1alpha1.TransactionPlan";
