+++ crates/core/component/ibc/Cargo.toml
@@ -26,6 +26,7 @@ penumbra-chain = { path = "../chain", default-features = false }
 penumbra-asset = { path = "../../../core/asset", default-features = false }
 penumbra-num = { path = "../../../core/num", default-features = false }
 penumbra-keys = { path = "../../../core/keys", default-features = false }
+penumbra-effecthash = { path = "../../../core/effecthash", default-features = false }
 
 # Penumbra dependencies
 ibc-types = { version = "0.10.0", default-features = false }
