+++ crates/core/component/shielded-pool/Cargo.toml
@@ -43,6 +43,7 @@ penumbra-chain = { path = "../chain", default-features = false }
 penumbra-asset = { path = "../../../core/asset", default-features = false }
 penumbra-num = { path = "../../../core/num", default-features = false }
 penumbra-keys = { path = "../../../core/keys", default-features = false }
+penumbra-effecthash = { path = "../../../core/effecthash", default-features = false }
 decaf377-ka = { path = "../../../crypto/decaf377-ka/" }
 decaf377-fmd = { path = "../../../crypto/decaf377-fmd/" }
