+++ crates/core/component/stake/Cargo.toml
@@ -45,6 +45,7 @@ penumbra-chain = { path = "../chain", default-features = false }
 penumbra-dao = { path = "../dao", default-features = false }
 penumbra-distributions = { path = "../distributions", default-features = false }
 penumbra-keys = { path = "../../../core/keys", default-features = false }
+penumbra-effecthash = { path = "../../../core/effecthash", default-features = false }
 penumbra-num = { path = "../../../core/num", default-features = false }
 penumbra-proof-params = { path = "../../../crypto/proof-params" }
 penumbra-proto = { path = "../../../proto" }
