+++ crates/core/component/stake/src/delegate.rs
@@ -1,4 +1,5 @@
 use penumbra_asset::{Balance, Value, STAKING_TOKEN_ASSET_ID};
+use penumbra_effecthash::{EffectHash, EffectingData};
 use penumbra_num::Amount;
 use penumbra_proto::{penumbra::core::component::stake::v1alpha1 as pb, DomainType};
 use serde::{Deserialize, Serialize};
@@ -24,6 +25,13 @@ pub struct Delegate {
     /// stateless verification that the transaction is internally consistent.
     /// TODO(erwan): make sure this is checked in tx validation
     pub delegation_amount: Amount,
+}
+
+impl EffectingData for Delegate {
+    fn effect_hash(&self) -> EffectHash {
+        // For delegations, the entire action is considered effecting data.
+        EffectHash::from_proto_effecting_data(&self.to_proto())
+    }
 }
 
 impl Delegate {
