+++ crates/core/effecthash/src/effect_hash.rs
@@ -1,4 +1,4 @@
-use penumbra_proto::{penumbra::core::effecthash::v1alpha1 as pb, DomainType};
+use penumbra_proto::{penumbra::core::effecthash::v1alpha1 as pb, DomainType, Message, Name};
 
 /// A hash of a transaction's _effecting data_, describing its effects on the
 /// chain state.
@@ -9,9 +9,32 @@ use penumbra_proto::{penumbra::core::effecthash::v1alpha1 as pb, DomainType};
 #[derive(Clone, Copy, Eq, PartialEq)]
 pub struct EffectHash(pub [u8; 64]);
 
+/// A helper function to create a BLAKE2b `State` instance given a variable-length personalization string.
+fn create_personalized_state(personalization: &str) -> blake2b_simd::State {
+    let mut state = blake2b_simd::State::new();
+
+    // The `TypeUrl` provided as a personalization string is variable length,
+    // so we first include the length in bytes as a fixed-length prefix.
+    let length = personalization.len() as u64;
+    state.update(&length.to_le_bytes());
+    state.update(personalization.as_bytes());
+
+    state
+}
+
 impl EffectHash {
     pub fn as_bytes(&self) -> &[u8; 64] {
         &self.0
+    }
+
+    /// A helper function to hash the data of a proto-encoded message, using
+    /// the variable-length `TypeUrl` of the corresponding domain type as a
+    /// personalization string.
+    pub fn from_proto_effecting_data<M: Message + Name>(message: &M) -> EffectHash {
+        let mut state = create_personalized_state(&M::type_url());
+        state.update(&message.encode_to_vec());
+
+        EffectHash(*state.finalize().as_array())
     }
 }
