+++ crates/core/transaction/src/plan/action.rs
@@ -1,12 +1,11 @@
 use crate::Action;
-use crate::EffectingData;
 use crate::WitnessData;
 use anyhow::{anyhow, Context, Result};
 use ark_ff::Zero;
 use decaf377::Fr;
 use penumbra_asset::Balance;
 use penumbra_dao::{DaoDeposit, DaoOutput, DaoSpend};
-use penumbra_effecthash::EffectHash;
+use penumbra_effecthash::{EffectHash, EffectingData};
 
 use penumbra_dex::{
     lp::{
